requirements.txt
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
pytz==2020.1
psycopg2==2.8.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-cors-headers==3.2.1

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
django-filter==2.2.0 # https://django-filter.readthedocs.io/en/latest/index.html

# guncorn + Sentry
gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
sentry-sdk==0.14.3  # https://github.com/getsentry/sentry-python
